refactoring to move json methods to element decision tree model to avoid having to extend another extension to update tree data
